Final version LpGBT Interface ...

See merge request cms_tk_ph2/Ph2_ACF!207

+++ Utils/linearFitter.h
@@ -38,7 +38,7 @@ void estimate_coef(std::vector<T> indep_var, std::vector<T> dep_var, double& B_1
 
     double zwischen_summe = 0;
 
-    for(int index = 0; index < indep_var.size(); ++index) { zwischen_summe += pow(B_1 * indep_var[index] + B_0 - dep_var[index], 2.); }
+    for(size_t index = 0; index < indep_var.size(); ++index) { zwischen_summe += pow(B_1 * indep_var[index] + B_0 - dep_var[index], 2.); }
     double DELTA = sqrt(zwischen_summe / (N - 2.));
 
     B_1_error = DELTA * sqrt(1. / (N * (mean_x_square - pow(mean_x, 2.))));
@@ -55,7 +55,7 @@ template <typename T>
 void estimate_coef(std::vector<T> indep_var, std::vector<T> dep_var, std::vector<T> errors, double& B_1, double& B_0, double& B_1_error, double& B_0_error)
 {
     std::vector<double> weights;
-    for(int index = 0; index < errors.size(); ++index)
+    for(size_t index = 0; index < errors.size(); ++index)
     {
         if(errors[index] == 0) { weights.push_back(0); }
         else
@@ -68,10 +68,10 @@ void estimate_coef(std::vector<T> indep_var, std::vector<T> dep_var, std::vector
     double mean_y      = std::inner_product(dep_var.begin(), dep_var.end(), weights.begin(), 0.0) / sum_weights;
 
     double mean_x_square = 0;
-    for(int index = 0; index < indep_var.size(); ++index) { mean_x_square += weights[index] * pow(indep_var[index], 2.); }
+    for(size_t index = 0; index < indep_var.size(); ++index) { mean_x_square += weights[index] * pow(indep_var[index], 2.); }
     mean_x_square /= sum_weights;
     double mean_x_times_y = 0;
-    for(int index = 0; index < indep_var.size(); ++index) { mean_x_times_y += weights[index] * indep_var[index] * dep_var[index]; }
+    for(size_t index = 0; index < indep_var.size(); ++index) { mean_x_times_y += weights[index] * indep_var[index] * dep_var[index]; }
     mean_x_times_y /= sum_weights;
     B_1 = (mean_x_times_y - mean_x * mean_y) / (mean_x_square - pow(mean_x, 2.));
     B_0 = (mean_x_square * mean_y - mean_x * mean_x_times_y) / (mean_x_square - pow(mean_x, 2.));
